server/requirements.txt
anyio==3.6.1
appdirs==1.4.3
asgiref==3.5.2
attrs==19.3.0
Automat==0.8.0
blinker==1.4
boto3==1.23.9
botocore==1.26.9
cached-property==1.5.1
certifi==2022.5.18.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
cloud-init==22.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
ec2-hibinit-agent==1.0.0
ecdsa==0.17.0
email-validator==1.2.1
entrypoints==0.3
fastapi==0.78.0
faunadb==4.2.0
filelock==3.0.12
future==0.18.2
greenlet==1.1.2
h11==0.13.0
h2==2.6.2
hibagent==1.0.1
hpack==3.0.0
httplib2==0.14.0
httptools==0.4.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==19.0.0
idna==3.3
importlib-metadata==1.5.0
incremental==16.10.1
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.1.1
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
orjson==3.6.8
pexpect==4.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyAudio==0.2.11
pydantic==1.9.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytube==12.1.0
PyYAML==5.3.1
requests==2.27.1
requests-unixsocket==0.2.0
rsa==4.8
s3transfer==0.5.2
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
sos==4.3
SQLAlchemy==1.4.36
sqlalchemy2-stubs==0.0.2a22
sqlmodel==0.0.6
ssh-import-id==5.10
starlette==0.19.1
systemd-python==234
texttable==1.6.2
Twisted==18.9.0
typing_extensions==4.2.0
ubuntu-advantage-tools==27.8
ufw==0.36
ujson==5.3.0
unattended-upgrades==0.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.0.17
wadllib==1.3.3
websocket-client==0.53.0
websockets==10.3
wget==3.2
zipp==1.0.0
zope.interface==4.7.1
